Authorization checks added to Comment Service.

+++ src/main/java/com/example/forum/services/CommentServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.forum.services;
 
+import com.example.forum.exceptions.AuthorizationException;
 import com.example.forum.models.Comment;
 import com.example.forum.models.Post;
 import com.example.forum.models.User;
@@ -10,6 +11,8 @@ import org.springframework.stereotype.Service;
 @Service
 public class CommentServiceImpl implements CommentService {
 
+    public static final String AUTHORIZATION_ERR = "Only admins or comment creators can modify this resource.";
+    public static final String USER_BLOCKED_ERR = "You are not allowed to create comments.";
     private final CommentRepository repository;
 
     @Autowired
@@ -19,6 +22,9 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public Comment create(Post post, Comment comment, User user) {
+        if(user.isBlocked()){
+            throw new AuthorizationException(USER_BLOCKED_ERR);
+        }
         comment.setCreatedBy(user);
         comment.setPost(post);
         repository.create(comment);
@@ -26,19 +32,28 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void update(Comment comment) {
+    public void update(Comment comment, User user) {
+        checkModifyPermissions(comment.getCommentId(), user);
         repository.update(comment);
-
 
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id, User user) {
+        checkModifyPermissions(id,user);
         repository.delete(id);
+
     }
 
     @Override
     public Comment getById(int id) {
         return repository.getById(id);
+    }
+
+    private void checkModifyPermissions(int id, User user) {
+       Comment comment = repository.getById(id);
+        if (!(user.isAdmin() || comment.getCreatedBy().equals(user))) {
+            throw new AuthorizationException(AUTHORIZATION_ERR);
+        }
     }
 }
